+++ package.json
@@ -28,6 +28,7 @@
     "react-native": "^0.55.4",
     "react-native-dropdown": "0.0.6",
     "react-native-elements": "^0.19.1",
+    "react-native-flip-view-next": "^1.0.0",
     "react-native-maps": "^0.21.0",
     "react-native-material-dropdown": "^0.11.1",
     "react-native-modal-dropdown": "^0.6.2",
