+++ src/screens/ListView.js
@@ -1,10 +1,11 @@
 import React, { Component } from "react";
+import { connect } from 'react-redux';
 import { Text, StyleSheet, View, Image, FlatList, TouchableOpacity, Dimensions } from "react-native";
 //import TEST_DATA from "../listingArray.json";
 import styles from '../style/listViewStyle';
 import Card from '../Components/Card';
 
-export default class ListView extends Component {
+class ListView extends Component {
   constructor(props) {
     super(props);
 
@@ -19,7 +20,8 @@ export default class ListView extends Component {
   render() {
     return (
       <View style={styles.container} onLayout = {this.onLayout.bind(this)}>
-         <FlatList
+        {this.props.filteredPropertiesList.length > 0 ? (
+          <FlatList
          style={styles.flatList}
           data={this.props.filteredPropertiesList}
           keyExtractor={item => item.id.toString()}
@@ -27,7 +29,15 @@ export default class ListView extends Component {
             <Card {...item} navigation={this.props.navigation} />
           )}
         />
+        ) : (
+          <Text style={{alignSelf: 'center', color: '#ccc'}}> No properties found! </Text>
+        )}
       </View>
     );
   }
 }
+const mapStateToProps = (state) => {
+  return {...state.searchOnMap}
+}
+
+export default connect(mapStateToProps)(ListView);
