+++ src/style/listViewStyle.js
@@ -2,6 +2,7 @@ import { StyleSheet } from "react-native";
 
 export default styles = StyleSheet.create({
   container: {
+    flex: 1,
     backgroundColor: "#ededed",
     justifyContent: "center",
     paddingTop: 20
