+++ src/style/propertyDetailStyle.js
@@ -2,7 +2,8 @@ import {StyleSheet } from 'react-native';
 
 export default styles = StyleSheet.create({
   componentContainer: {
-    flex: 1
+    flex: 1,
+    backgroundColor: '#f7f7f7'
   },
   carouselContainer: {
     flex: 1,
